fix(homefiles): only copy over config file if it exists

+++ modules/homefiles/homefiles.sh
@@ -43,7 +43,9 @@ if [[ ${#ADD_HOMEFILES[@]} -gt 0 ]]; then
 	for entry in "${ADD_HOMEFILES[@]}"; do
 		cp -r $CONFIG_DIRECTORY/homefiles/$entry /usr/share/bluebuild/homefiles/$entry
 		mkdir -p /usr/etc/bluebuild/homefiles/$entry
-		cp -r $CONFIG_DIRECTORY/homefiles/$entry/chezmoi.toml /usr/etc/bluebuild/homefiles/$entry/chezmoi.toml
+		if [ -e $CONFIG_DIRECTORY/homefiles/$entry/chezmoi.toml ]; then
+			cp -r $CONFIG_DIRECTORY/homefiles/$entry/chezmoi.toml /usr/etc/bluebuild/homefiles/$entry/chezmoi.toml
+		fi
 		chezmoi apply --destination /usr/etc/skel/ --source /usr/share/bluebuild/homefiles/$entry --force
 	done
 else
